Change CCL input precision from int8 to int64 to align with compiler (#1003)
This PR is for changing the precision of the CCL input from int8 to
int64 to be aligned with the compiler for release 1.22. This change is
for addressing the JIRA tickets raised when working with CCL enabled in
new sdk versions.

+++ QEfficient/transformers/models/llama4/modeling_llama4.py
@@ -1253,7 +1253,7 @@ class QEffLlama4ForConditionalGeneration(Llama4ForConditionalGeneration):
             lang_inputs["batch_index"] = torch.arange(bs).view(bs, 1)
 
         if comp_ctx_lengths is not None:
-            lang_inputs["comp_ctx_lengths"] = torch.randint(0, 100, (40,), dtype=torch.int8)
+            lang_inputs["comp_ctx_lengths"] = torch.randint(0, 100, (40,), dtype=torch.int64)
 
         inputs = {}
         if kv_offload:
